requirements.txt
python-dotenv
requests
openai
moviepy
gtts
google-api-python-client
google-auth-httplib2
google-auth-oauthlib 